populator uses stream monad

+++ src/main/scala/traindelays/scripts/PopulateScheduleTable.scala
@@ -8,6 +8,7 @@ import traindelays.networkrail.NetworkRailClient
 import traindelays.networkrail.db._
 import traindelays.networkrail.scheduledata._
 import cats.syntax.flatMap._
+import fs2.Stream
 
 trait PopulateScheduleTable extends StrictLogging {
 
@@ -19,34 +20,38 @@ trait PopulateScheduleTable extends StrictLogging {
     val scheduleTableSecondary = ScheduleSecondaryTable(db, config.networkRailConfig.scheduleData.memoizeFor)
     val associationTable       = AssociationTable(db, config.networkRailConfig.scheduleData.memoizeFor)
 
-    fs2.Stream.eval {
-      for {
-        _      <- IO(logger.info(s"Starting population of schedule table. Flush first set to $flushFirst"))
-        client <- Http1Client[IO]()
-        networkRailClient  = NetworkRailClient(config.networkRailConfig, client)
-        scheduleDataReader = ScheduleDataReader(config.networkRailConfig.scheduleData.tmpUnzipLocation)
-        _ <- networkRailClient.deleteTmpFiles()
-        _ <- IO(logger.info("Downloading schedule data"))
-        _ <- downloadScheduleData(networkRailClient).compile.drain
-        _ <- IO(logger.info("Unpacking schedule data"))
-        _ <- networkRailClient.unpackScheduleData.compile.drain
-        _ <- if (flushFirst)
+    for {
+      _      <- Stream.eval(IO(logger.info(s"Starting population of schedule table. Flush first set to $flushFirst")))
+      client <- Http1Client.stream[IO]()
+      networkRailClient  = NetworkRailClient(config.networkRailConfig, client)
+      scheduleDataReader = ScheduleDataReader(config.networkRailConfig.scheduleData.tmpUnzipLocation)
+      _ <- Stream.eval(networkRailClient.deleteTmpFiles())
+      _ <- Stream.eval(IO(logger.info("Downloading schedule data")))
+      _ <- downloadScheduleData(networkRailClient)
+      _ <- Stream.eval(IO(logger.info("Unpacking schedule data")))
+      _ <- networkRailClient.unpackScheduleData
+      _ <- Stream.eval[IO, Unit] {
+        if (flushFirst)
           IO(logger.info("Deleting all records from Schedule Table Primary")) >> scheduleTablePrimary.deleteAllRecords()
         else IO.unit
-        _ <- if (flushFirst)
+      }
+      _ <- Stream.eval[IO, Unit] {
+        if (flushFirst)
           IO(logger.info("Deleting all records from Association Table")) >> associationTable.deleteAllRecords()
         else IO.unit
-        _ <- if (flushFirst) IO(logger.info("Deleting all records from Stanox Table")) >> stanoxTable.deleteAllRecords()
+      }
+      _ <- Stream.eval[IO, Unit] {
+        if (flushFirst) IO(logger.info("Deleting all records from Stanox Table")) >> stanoxTable.deleteAllRecords()
         else IO.unit
-        _ <- IO(logger.info("Writing schedule data records"))
-        _ <- writeScheduleDataRecords(stanoxTable,
-                                      scheduleTablePrimary,
-                                      scheduleTableSecondary,
-                                      associationTable,
-                                      scheduleDataReader).compile.drain
-        _ <- IO(logger.info("Schedule Table population complete"))
-      } yield ()
-    }
+      }
+      _ <- Stream.eval(IO(logger.info("Writing schedule data records")))
+      _ <- writeScheduleDataRecords(stanoxTable,
+                                    scheduleTablePrimary,
+                                    scheduleTableSecondary,
+                                    associationTable,
+                                    scheduleDataReader)
+      _ <- Stream.eval(IO(logger.info("Schedule Table population complete")))
+    } yield ()
   }
   protected def downloadScheduleData(networkRailClient: NetworkRailClient): fs2.Stream[IO, Unit]
 }
